// backend/src/types/employee.ts
export interface Employee {
  id: string;
  employeeCode: string;
  firstName: string;
  lastName: string;
  email: string;
  country: string;
  department: string;
  roleTitle: string;
  level: string;
  currency: string;
  salaryAmount: number;
  hireDate: string;
  status: string;
  createdAt: string;
  updatedAt: string;
}